% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEPlotStacking.R
\name{ALEPlotStacking}
\alias{ALEPlotStacking}
\title{Accumulated Local Effects (ALE) Plots of a
stacking of predictive distribution model.}
\usage{
ALEPlotStacking(X, X.model, J, which.alpha, K = 40, plot_output = TRUE,
  confidence.intervals = TRUE, ylim = NULL)
}
\arguments{
\item{X}{Data frame with the predictor variables used in the formula,
not including the response.}

\item{X.model}{The fitted model.}

\item{J}{A numeric scalar indicating the column of the predictor
in the model matrix for which the ALE plot will be calculated.
At the moment only effects of single covariates are implemented (no interactions)}

\item{which.alpha}{A numeric scalar indicating the category for which the
effect will be plotted.}

\item{K}{Number of intervals into which the predictor range is divided
when calculating the ALE plot effects}

\item{confidence.intervals}{A logical value. If TRUE,
approximate posterior credible intervals based on the Delta method are plotted.
Intervals are available only for numeric variables, not for factor variables.}
}
\value{
The ALE plot of the covariate effects over each category/expert.
}
\description{
Accumulated Local Effects (ALE) Plots of a
stacking of predictive distribution model.
}
