% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_bspline.R
\name{basis_bspline}
\alias{basis_bspline}
\title{Get B-spline design matrix constructor}
\usage{
basis_bspline(k, m, lim)
}
\arguments{
\item{k}{number of knots.}

\item{m}{order of the B-spline basis and penalty. For instance \code{m = 3} creates
a 3rd degree basis with a second order penalty.}

\item{lim}{range over which the basis is created.}
}
\value{
A function that take arguments \code{x} and \code{deriv} and returns a P-Spline model matrix and its first
\code{deriv} derivatives w.r.t. \code{x}.
}
\description{
Get B-spline design matrix constructor
}
\examples{
library(gamFactory)

tmp <- basis_bspline(k = 6, 
                     m = 3, 
                     lim = c(-1, 1)
                     )$evalX(x = seq(-0.9, 0.9, 0.01), deriv = 3)
                     
par(mfrow = c(2, 2)) 
matplot(tmp$X0, type = 'l') # Design matrix
matplot(tmp$X1, type = 'l') # 1st derivative w.r.t. x 
matplot(tmp$X2, type = 'l') # 2nd derivative w.r.t. x 
matplot(tmp$X3, type = 'l') # 3rd derivative w.r.t. x

}
