% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_family.R
\name{build_family}
\alias{build_family}
\title{Building new GAM families for mgcv::gam}
\usage{
build_family(bundle)
}
\arguments{
\item{np}{XXX.}
}
\description{
XXX.
}
\examples{
fam <- build_family(bundle_fam_gaussian())
library(MASS)
b <- gam(list(accel~s(times,k=20,bs="ad"),~s(times)),
         data=mcycle,family=fam)
summary(b) 
plot(b,pages=1,scale=0)

}
