% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMultiLP.R
\name{buildMultiLP}
\alias{buildMultiLP}
\title{Build a linear predictor}
\usage{
buildMultiLP(eff, iel, iec)
}
\arguments{
\item{iel}{vectors indicating to which linear predictor each component belongs to.}

\item{iec}{list of vectors indicating which regression coefficients belong to each component.}
}
\description{
Build a linear predictor
}
