% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMultipleLP.R
\name{buildMultipleLP}
\alias{buildMultipleLP}
\title{Build a linear predictor}
\usage{
buildMultipleLP(o, iel, iec)
}
\arguments{
\item{iel}{vectors indicating to which linear predictor each component belongs to.}

\item{iec}{list of vectors indicating which regression coefficients belong to each component.}
}
\description{
Build a linear predictor
}
