% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_fam_multinomial.R
\name{bundle_fam_multinomial}
\alias{bundle_fam_multinomial}
\title{Bundle for multinomial logistic regression model}
\usage{
bundle_fam_multinomial(K)
}
\arguments{
\item{K}{number of linear predictors, i.e. number of categories minus one.}
}
\description{
Ported from \code{\link[mgcv:multinom]{mgcv::multinom}}. There are \code{K} linear predictors, one
per non-reference category, each giving the log-odds relative to
category \code{0} (the reference category, which has no linear predictor
of its own).
}
