% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_fam_poisson.R
\name{bundle_fam_poisson}
\alias{bundle_fam_poisson}
\title{Bundle for the Poisson regression model}
\usage{
bundle_fam_poisson()
}
\description{
Bundle for a Poisson GAM with a single linear predictor for the rate
\code{mu}.
}
