% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_fam_twlss.R
\name{bundle_fam_twlss}
\alias{bundle_fam_twlss}
\title{Bundle for the Tweedie location-scale-shape model}
\usage{
bundle_fam_twlss(a = 1.01, b = 1.99)
}
\arguments{
\item{a}{lower limit parameter (> 1) used in the definition of the Tweedie power
parameter \code{p} from \code{theta}. See \code{mgcv::ldTweedie}.}

\item{b}{upper limit parameter (< 2) used in the definition of the Tweedie power
parameter \code{p} from \code{theta}. See \code{mgcv::ldTweedie}.}
}
\description{
Ported from \code{\link[mgcv:twlss]{mgcv::twlss}}. The three linear predictors correspond,
in order, to the mean \code{mu}, \code{theta} (which determines the
Tweedie power parameter via \code{p = (a + b * exp(theta)) / (1 + exp(theta))})
and \code{rho = log(phi)}, the log dispersion. Only 1st and 2nd order
derivatives are available (as in \code{mgcv::twlss}, whose
\code{available.derivs} is 0).
}
