% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_ztnbinom2.R
\name{bundle_ztnbinom2}
\alias{bundle_ztnbinom2}
\title{Bundle for zero-truncated negative binomial model}
\usage{
bundle_ztnbinom2(theta = NULL)
}
\arguments{
\item{theta}{A positive supplied \code{theta} is treated as a fixed known parameter, otherwise it is estimated.}
}
\description{
Bundle for zero-truncated negative binomial model
}
