% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_ztpoisson.R
\name{bundle_ztpoisson}
\alias{bundle_ztpoisson}
\title{Bundle for zero-truncated Poisson regression model}
\usage{
bundle_ztpoisson()
}
\description{
Bundle for zero-truncated Poisson regression model
}
