% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundleMARKOV.R
\docType{data}
\name{bundleMARKOV}
\alias{bundleMARKOV}
\title{Ingredients for two states Markov model}
\format{An object of class \code{list} of length 7.}
\usage{
bundleMARKOV
}
\description{
XXX.
}
\examples{
library(gamFactory)

# Simulate some data
n <- 2000
myDat <- data.frame(x = sort(runif(n, -1, 1)))
p1 <- pnorm(myDat$x, 0, 0.7)[-1]
p2 <- pnorm(-myDat$x, 0, 0.7)[-1]

y0 <- numeric(n)
y0[1] <- 1
for(ii in 2:n){
  if(y0[ii-1] == 0) y0[ii] <- rbinom(1, 1, 1 - p1[ii])
  if(y0[ii-1] == 1) y0[ii] <- rbinom(1, 1, 1 - p1[ii])
}

myDat$y <-  y0
plot(myDat)

# Create GPD family and fit it
library(mgcViz)
myMRK <- buildGAMLSS( bundleMARKOV )
fit <- gamV(list(y ~ s(x), ~ s(x)), data = myDat, family = myMRK)
print(plot(fit), pages = 1)

# Compare true (black) and estimated (red) quantiles
qu <- 0.95
xseq <- seq(-1, 1, length.out = 1000)
plot(myDat, col = "grey")
lines(xseq, qexp(qu, 1 / (0.3 + abs(xseq)^2)), col = 2)
lines(xseq, fit$family$qf(qu,
                          mu = predict(fit, newdata = data.frame(x = xseq),
                                       type = "response")),
      col = 3)
#' 
}
\keyword{datasets}
