% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_deriv_auto.R
\name{check_deriv_auto}
\alias{check_deriv_auto}
\title{Checking derivatives by finited differences}
\usage{
check_deriv_auto(
  np,
  simfun,
  obj,
  ord = 1,
  trans = identity,
  ...,
  fun = "derobj"
)
}
\arguments{
\item{np}{number of parameter vectors to be simulated.}

\item{simfun}{function to simulate parameter vectors.}

\item{obj}{for instance the output of createSH}

\item{ord}{the order to derivatives to be checked.}

\item{trans}{transformation to be applied to the derivatives}

\item{...}{arguments passed to \code{obj$initialize}.}

\item{fun}{the name of the slot of \code{obj} containing the derivative function.}
}
\description{
Function for comparing exact and numerical derivatives.
}
