% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_g_derivatives_fd.R
\name{check_g_derivatives_fd}
\alias{check_g_derivatives_fd}
\alias{.get_nested_index}
\title{Function used in build_family_nl to check the correctness of g1 and g2}
\usage{
.get_nested_index(olp)
}
\description{
only suitable for "si_nexp" effect
}
