% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDeriv.R
\name{checkDeriv}
\alias{checkDeriv}
\title{Comparing analytic with numerical finite differences}
\usage{
checkDeriv(obj, param, ord = 1)
}
\arguments{
\item{obj}{the function to be differentiated.}

\item{param}{vector of parameters at which we should calculate the derivatives.}

\item{ord}{the order to derivatives to be checked.}
}
\description{
Function for getting higher derivatives by finited differences.
}
