% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDerivAuto.R
\name{checkDerivAuto}
\alias{checkDerivAuto}
\title{Checking derivatives by finited differences}
\usage{
checkDerivAuto(
  np,
  parSim,
  obj,
  ord = 1,
  trans = identity,
  ...,
  derFunNam = "derObj"
)
}
\arguments{
\item{np}{number of parameter vectors to be simulated.}

\item{parSim}{function to simulate parameter vectors.}

\item{obj}{for instance the output of createSH}

\item{ord}{the order to derivatives to be checked.}

\item{trans}{trasformation to be applied to the derivatives}

\item{...}{arguments passed to \code{obj$initialize}.}

\item{derFunNam}{the name of the slot of \code{obj} containing the derivative function.}
}
\description{
Function for comparing exact and numerical derivatives.
}
