% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDerivStackConvexOld.R
\name{convertDerivStack}
\alias{convertDerivStack}
\title{Calculating derivatives of log-likelihood wrt regression coefficients}
\usage{
convertDerivStack(param, theta, X, Z, le, lee, leee = 0, d1b = 0, deriv = 0)
}
\arguments{
\item{param}{regression coefficients}

\item{theta}{vector of other parameters of the response variable distribution (to be included in beta to be estimated by ML)}

\item{X}{nxK model matrix of the convex combination}

\item{Z}{nxp model matrix}

\item{le}{nx1 vector of 1st derivatives of the log-likelihood with respect to eta}

\item{lee}{nx1 vector of 2nd derivatives of the log-likelihood with respect to eta}

\item{leee}{nx1 vector of 3rd derivatives of the log-likelihood with respect to eta}

\item{d1b}{(p(K-1))xm matrix of derivatives of the Hessian with respect to the smoothing parameters (at the moment only one smoothing parameter)}

\item{deriv}{0: just grad and Hess, 1: first deriv of Hess}
}
\description{
Old, it implements stack effects with convex combination,
it only provides derivatives with respect to the regression coefficients and
not with respect to the other parameters of the response distribution.
Intended for use with stack effects.
Converts derivatives of the log-likelihood with respect to
the non linear predictor eta into derivatives with respect to the
regression coefficients and smoothing parameters
}
