% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDirPrior.R
\name{createDirPrior}
\alias{createDirPrior}
\title{Prior on direction of single index vector}
\usage{
createDirPrior(r = NULL, alpha0 = NULL, alphaRho = NULL)
}
\description{
This is specific to single index effects.
}
