% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGaussian.R
\name{createGaussian}
\alias{createGaussian}
\title{Creating a Gaussian distribution}
\usage{
createGaussian(y = NULL)
}
\arguments{
\item{y}{a vector of observations.}
}
\description{
This function creates an object representing a Gaussian distribution
}
