% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGPD.R
\name{createGPDs}
\alias{createGPDs}
\alias{createGPD}
\title{Creating a Generalized Pareto Distribution}
\usage{
createGPD(y = NULL)
}
\arguments{
\item{y}{a vector of observations.}
}
\description{
This function creates an object representing a GPD.
}
