% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSCEffect.R
\name{createSCEffect}
\alias{createSCEffect}
\title{Creating a smooth constrained effect}
\usage{
createSCEffect(x = NULL)
}
\description{
This function creates an object representing a smooth constrained effect.
At the moment, there is one effect that is constrained to be nondecreasing
}
