% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSH.R
\name{createSH}
\alias{createSH}
\title{Creating a Sinh-Arsinh (shash) distribution}
\usage{
createSH(y = NULL)
}
\arguments{
\item{y}{a vector of observations.}
}
\description{
This function creates an object representing a SHASH distribution
}
