% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarPrior.R
\name{createVarPrior}
\alias{createVarPrior}
\title{Prior on variance of single index model}
\usage{
createVarPrior(X = NULL, v = NULL, alpha0 = NULL, rhoAlpha = NULL)
}
\description{
This is specific to single index effects.
}
