% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/der_multiLP.R
\name{der.multiLP}
\alias{der.multiLP}
\title{Derivatives of log-likelihood with multiple linear predictor}
\usage{
\method{der}{multiLP}(o, llk, deriv = 1, param = NULL)
}
\description{
Derivatives of log-likelihood with multiple linear predictor
}
