% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{der}
\alias{der}
\title{Generic derivative function}
\usage{
der(o, ...)
}
\arguments{
\item{o}{object we want to get the derivatives of.}

\item{...}{arguments to be passed to methods.}
}
\description{
Generic derivative function
}
