% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_DhDr.R
\name{DHessDrho.easy}
\alias{DHessDrho.easy}
\title{Derivative of Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{easy}(o, llk, DbDr, index)
}
\description{
Derivative of Hessian w.r.t smoothing parameters
}
