% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_expsmooth.R
\name{DHessDrho.expsmooth}
\alias{DHessDrho.expsmooth}
\title{Derivative of exp smooth Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{expsmooth}(o, llk, DbDr)
}
\description{
Derivative of exp smooth Hessian w.r.t smoothing parameters
}
