% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_inter_linear.R
\name{DHessDrho.inter_linear}
\alias{DHessDrho.inter_linear}
\title{❗️Maybe we can remove this function if we return "si" class in the end of "eff_inter"}
\usage{
\method{DHessDrho}{inter_linear}(o, llk, DbDr)
}
\description{
Derivative of inter_linear Hessian w.r.t smoothing parameters
}
