% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_linearPredictor.R
\name{DHessDrho.linearPredictor}
\alias{DHessDrho.linearPredictor}
\title{Derivative of Hessian w.r.t smoothing parameters}
\usage{
DHessDrho.linearPredictor(o, llk, DbDr)
}
\description{
Derivative of Hessian w.r.t smoothing parameters
}
