% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_linpreds.R
\name{DHessDrho.linpreds}
\alias{DHessDrho.linpreds}
\title{Derivative of Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{linpreds}(o, llk, DbDr)
}
\description{
Derivative of Hessian w.r.t smoothing parameters
}
