% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_mgks.R
\name{DHessDrho.mgks}
\alias{DHessDrho.mgks}
\title{Derivative of MGKS Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{mgks}(o, llk, DbDr)
}
\description{
Derivative of MGKS Hessian w.r.t smoothing parameters
}
