% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_multiLP.R
\name{DHessDrho.multiLP}
\alias{DHessDrho.multiLP}
\title{Derivative of Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{multiLP}(o, llk, DbDr)
}
\description{
Derivative of Hessian w.r.t smoothing parameters
}
