% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_nexpsm.R
\name{DHessDrho.nexpsm}
\alias{DHessDrho.nexpsm}
\title{Derivative of exp smooth Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{nexpsm}(o, llk, DbDr)
}
\description{
Derivative of exp smooth Hessian w.r.t smoothing parameters
}
