% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_si.R
\name{DHessDrho.si}
\alias{DHessDrho.si}
\title{Derivative of single index Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{si}(o, llk, DbDr)
}
\description{
Derivative of single index Hessian w.r.t smoothing parameters
}
