% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_si_nexp.R
\name{DHessDrho.si_nexpsm}
\alias{DHessDrho.si_nexpsm}
\title{Derivative of si_nexpsm Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{si_nexpsm}(o, llk, DbDr)
}
\description{
Derivative of si_nexpsm Hessian w.r.t smoothing parameters
}
