% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_si_posi.R
\name{DHessDrho.si_posi}
\alias{DHessDrho.si_posi}
\title{Derivative of si_posi Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{si_posi}(o, llk, DbDr)
}
\description{
Derivative of si_posi Hessian w.r.t smoothing parameters
}
