% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_singleIndex.R
\name{DHessDrho.singleIndex}
\alias{DHessDrho.singleIndex}
\title{Derivative of single index Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{singleIndex}(o, llk, DbDr)
}
\description{
Derivative of single index Hessian w.r.t smoothing parameters
}
