% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho.stand.R
\name{DHessDrho.stand}
\alias{DHessDrho.stand}
\title{Derivative of standard Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{stand}(o, llk, DbDr)
}
\description{
Derivative of standard Hessian w.r.t smoothing parameters
}
