% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_standard.R
\name{DHessDrho.standard}
\alias{DHessDrho.standard}
\title{Derivative of standard Hessian w.r.t smoothing parameters}
\usage{
\method{DHessDrho}{standard}(o, llk, DbDr)
}
\description{
Derivative of standard Hessian w.r.t smoothing parameters
}
