% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DllkDbeta_dlinear.R
\name{DllkDbeta.inter_dlinear}
\alias{DllkDbeta.inter_dlinear}
\title{Derivatives of nested interaction (double single-index) effects}
\usage{
\method{DllkDbeta}{inter_dlinear}(o, llk, deriv = 1, param = NULL)
}
\description{
Derivatives of nested interaction (double single-index) effects
}
