% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DllkDbeta_inter_ee.R
\name{DllkDbeta.inter_ee}
\alias{DllkDbeta.inter_ee}
\title{Derivatives of the exp(x1) / exp(x2) interaction effect}
\usage{
\method{DllkDbeta}{inter_ee}(o, llk, deriv = 1, param = NULL)
}
\description{
Derivatives of the exp(x1) / exp(x2) interaction effect
}
