% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DllkDbeta_inter_le.R
\name{DllkDbeta.inter_le}
\alias{DllkDbeta.inter_le}
\title{Derivatives of the si(x) / exp(x) interaction effect}
\usage{
\method{DllkDbeta}{inter_le}(o, llk, deriv = 1, param = NULL)
}
\description{
Derivatives of the si(x) / exp(x) interaction effect
}
