% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DllkDbeta_linpreds.R
\name{DllkDbeta.linpreds}
\alias{DllkDbeta.linpreds}
\title{Derivatives of log-likelihood w.r.t. coefs of multiple linear predictors}
\usage{
\method{DllkDbeta}{linpreds}(o, llk, deriv = 1, param = NULL)
}
\description{
Derivatives of log-likelihood w.r.t. coefs of multiple linear predictors
}
