% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DllkDbeta_multiLP.R
\name{DllkDbeta.multiLP}
\alias{DllkDbeta.multiLP}
\title{Derivatives of log-likelihood w.r.t. coefs of multiple linear predictors}
\usage{
\method{DllkDbeta}{multiLP}(o, llk, deriv = 1, param = NULL)
}
\description{
Derivatives of log-likelihood w.r.t. coefs of multiple linear predictors
}
