% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{DllkDbeta}
\alias{DllkDbeta}
\title{Generic derivative of log-lik w.r.t. regression coefficients}
\usage{
DllkDbeta(o, ...)
}
\arguments{
\item{o}{object we want to get the derivatives of.}

\item{...}{arguments to be passed to methods.}
}
\description{
Generic derivative of log-lik w.r.t. regression coefficients
}
