% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DllkDbeta_singleIndex.R
\name{DllkDbeta.singleIndex}
\alias{DllkDbeta.singleIndex}
\title{Derivatives of single index effects}
\usage{
\method{DllkDbeta}{singleIndex}(o, llk, deriv = 1, param = NULL)
}
\description{
Derivatives of single index effects
}
