% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DllkDbeta.stand.R
\name{DllkDbeta.stand}
\alias{DllkDbeta.stand}
\title{Derivatives of log-lik w.r.t. coeffs of a standard effect}
\usage{
\method{DllkDbeta}{stand}(o, llk, deriv = 1, param = NULL)
}
\description{
Derivatives of log-lik w.r.t. coeffs of a standard effect
}
