% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DllkDbeta_standard.R
\name{DllkDbeta.standard}
\alias{DllkDbeta.standard}
\title{Derivatives of log-lik w.r.t. coeffs of a standard effect}
\usage{
\method{DllkDbeta}{standard}(o, llk, deriv = 1, param = NULL)
}
\description{
Derivatives of log-lik w.r.t. coeffs of a standard effect
}
