% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_inter.R
\name{eff_inter}
\alias{eff_inter}
\title{Build interaction effect}
\usage{
eff_inter(Xi, t, basis, a0 = NULL)
}
\arguments{
\item{Xi}{matrix to be projected via single index vector \code{alpha}.}

\item{basis}{function which takes \code{si = Xi\%*\%alpha} as input and returns model
matrix and its derivatives w.r.t. \code{si}.}
}
\description{
Build interaction effect
}
