% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_inter_dlinear.R
\name{eff_inter_dlinear}
\alias{eff_inter_dlinear}
\title{Build interaction_dlinear effect}
\usage{
eff_inter_dlinear(Xi, basis, a0 = NULL)
}
\arguments{
\item{Xi}{A combined matrix, with attributes "na1" and "na2".}

\item{basis}{function which takes z1 and z2 as inputs and returns the model matrix.}

\item{a0}{A numeric vector (concatenated, length na1 + na2) of initialization shifts.}
}
\description{
Build interaction_dlinear effect
}
