% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_inter_nexp.R
\name{eff_inter_nexp}
\alias{eff_inter_nexp}
\title{Build interaction_linear effect}
\usage{
eff_inter_nexp(y, Xi, t, basis, x0 = NULL, times = NULL)
}
\arguments{
\item{Xi}{matrix to be projected via single index vector \code{alpha}.}

\item{basis}{function which takes \code{si = Xi\%*\%alpha} as input and returns model
matrix and its derivatives w.r.t. \code{si}.}
}
\description{
Build interaction_linear effect
}
