% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_multinomial.R
\name{fam_multinomial}
\alias{fam_multinomial}
\title{The multinomial family}
\usage{
fam_multinomial(K, link = NULL)
}
\arguments{
\item{K}{number of linear predictors, i.e. number of categories minus one.}
}
\description{
Ported from \code{\link[mgcv:multinom]{mgcv::multinom}}, see the corresponding documentation for details.
}
\examples{
library(gamFactory)

set.seed(0)
n <- 500; K <- 2
x <- runif(n)
eta <- cbind(2 * x - 1, 1 - 2 * x)
p <- cbind(1, exp(eta)); p <- p / rowSums(p)
y <- apply(p, 1, function(pr) sample(0:K, 1, prob = pr))
dat <- data.frame(y = y, x = x)

fit <- gam(list(y ~ s(x), ~ s(x)), data = dat, family = fam_multinomial(K = K))
plot(fit, pages = 1)

}
