% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdDeriv.R
\name{fdDeriv}
\alias{fdDeriv}
\title{Vector-valued finited differences}
\usage{
fdDeriv(obj, param, ord = 1)
}
\arguments{
\item{obj}{the function to be differentiated.}

\item{param}{vector of parameters at which we should calculate the derivatives.}

\item{ord}{the order to derivatives to be checked.}
}
\description{
Function for getting higher derivatives by finited differences.
}
