% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_family_link.R
\name{fix_family_link}
\alias{fix_family_link}
\title{Preparing response family function}
\usage{
fix_family_link(fam)
}
\arguments{
\item{fam}{a family, such as the output of \code{fam_gaussian}.}
}
\description{
Extension of \code{mgcv::fix.family.link} including additional links.
It purpose is to add the necessary link derivatives to a family object,
if they are not already present.
}
\examples{
fix_family_link(fam_gaussian())$link

}
