% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixFamilyLink.R
\name{fixFamilyLink}
\alias{fixFamilyLink}
\title{Preparing response family function}
\usage{
fixFamilyLink(fam)
}
\arguments{
\item{np}{XXX.}
}
\description{
Extension of \code{mgcv::fix.family.link} including additional links.
}
