% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenSymMat.R
\name{flattenSymMat}
\alias{flattenSymMat}
\title{Transform symmetric matrix to vector}
\usage{
flattenSymMat(H)
}
\arguments{
\item{H}{a symmetric matrix.}
}
\value{
A vector of unique elements of H, by row.
}
\description{
The input is a symmetric matrix, the output is a vector of its unique elements by row.
}
