% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamFactory.R
\docType{_PACKAGE}
\name{gamFactory}
\alias{gamFactory}
\title{gamFactory: Tools for building new GAM models in mgcv.}
\description{
gamFactory: Tools for building new GAM models in mgcv.
}
\section{gamFactory functions}{

The gamFactory functions ...
}

