% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamnl.R
\name{gamnl}
\alias{gamnl}
\title{Fit GAM model with non-linear effects}
\usage{
gamnl(formula, family = gaulss(), data = list(), ...)
}
\arguments{
\item{formula, family, data}{same arguments as in \link[mgcv:gam]{mgcv::gam}.}

\item{...}{further arguments to be passed to \link[mgcv:gam]{mgcv::gam}.}
}
\description{
Fit GAM model with non-linear effects
}
