% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixFamilyLink.R, R/getStats.R, R/makeLink.R
\name{fixFamilyLink}
\alias{fixFamilyLink}
\alias{getStats}
\alias{makeLink}
\title{Preparing response family function}
\usage{
fixFamilyLink(fam)

getStats(link, okLinks, np, nam)

makeLink(link)
}
\arguments{
\item{link}{a character string indicating the link function to be used.}

\item{np}{XXX.}
}
\description{
Extension of \code{mgcv::fix.family.link} including additional links.

XXX.

Extension of \code{mgcv::make.link} including additional links.
}
\examples{
library(gamFactory)
getStats(link = "logitab(1.2, 4)", "logitab(3, 5)", 1, "gaussian")
getStats(link = "loga(1.2)", "loga", 1, "gaussian")

}
