% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpreds.R
\name{linpreds}
\alias{linpreds}
\title{Build linear predictors}
\usage{
linpreds(eff, iel, iec)
}
\arguments{
\item{eff}{list of effects, see e.g. \link{eff_stand} or \link{eff_nest}.}

\item{iel}{vectors indicating to which linear predictor each effect belongs to.}

\item{iec}{list of vectors indicating which regression coefficients belong to each effect.}
}
\description{
Build linear predictors
}
