% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_binomial.R
\name{llk_binomial}
\alias{llk_binomial}
\title{Log-likelihood of a Binomial model}
\usage{
llk_binomial(y, param, n, deriv = 0, ...)
}
\arguments{
\item{y}{a vector of observed counts (number of successes out of \code{n} trials).}

\item{param}{a matrix (or list) with 1 column (element), the success probability
\code{mu}.}

\item{n}{number of trials, either a single value (recycled for every observation) or
a vector of the same length as \code{y}.}

\item{deriv}{integer between 0 and 4 indicating the maximum derivative order to
return: 0 only returns \code{d0} (the log-density itself), while 1-4
additionally return \code{d1}-\code{d4}.}
}
\description{
Log-likelihood of the Binomial distribution, and its derivatives with
respect to \code{param = mu}, the success probability (\code{y} counts
successes out of \code{n} trials, both on the response, natural scale).
Returned in the list format used by \code{[gamFactory::llk_gaussian]}
and friends.
}
\examples{
library(gamFactory)
nobs <- 10
n <- 50
y <- rbinom(nobs, size = n, prob = 0.3)
param <- c(0.4)

# Derivatives of Binomial log-lik up to order 4
llk_binomial(y = y, param = param, n = n, deriv = 4)

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(
  "d0" = function(param){
    sum(llk_binomial(y = y, param = param, n = n, deriv = 0)$d0)
  },
  "d1" = function(param){
    colSums(do.call("cbind", llk_binomial(y = y, param = param, n = n, deriv = 1)$d1))

  },
  "d2" = function(param){
    colSums(do.call("cbind", llk_binomial(y = y, param = param, n = n, deriv = 2)$d2))

  },
  "d3" = function(param){
    colSums(do.call("cbind", llk_binomial(y = y, param = param, n = n, deriv = 3)$d3))
  },
  "d4" = function(param){
    colSums(do.call("cbind", llk_binomial(y = y, param = param, n = n, deriv = 4)$d4))
  })

check_deriv(obj = obj, param = param, ord = 1:4)

}
