% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_gammals.R
\name{llk_gammals}
\alias{llk_gammals}
\title{Log-likelihood of the Gamma location-scale distribution}
\usage{
llk_gammals(y, param, deriv = 0, ...)
}
\arguments{
\item{y}{a vector of observations (positive).}

\item{param}{a matrix (or list) with 2 columns (elements), containing \code{mu} and
\code{th}, in this order.}

\item{deriv}{non-negating positive integer between 0 and 4, indicating the derivative level requires.}
}
\description{
Log-likelihood of the Gamma location-scale model, and its derivatives
with respect to \code{param = cbind(mu, th)}, where \code{mu = log(mean)}
(note: despite the "identity" link name used by \code{mgcv::gammals} for
this parameter, it is internally the log of the mean) and
\code{th = log(phi)} is the log dispersion (\code{phi = 1 / shape}).
Ported directly from the derivative system used internally by
\link[mgcv:gammals]{mgcv::gammals}, but re-expressed as derivatives w.r.t. \code{param}
and returned in the list format used by \link{llk_gaussian}
and friends. See \link[mgcv:gammals]{mgcv::gammals} for background on the
parametrization.
}
\examples{
library(gamFactory)

n <- 10
param <- cbind(mu = log(runif(n, 1, 3)), th = rnorm(n, 0, 0.3))
y <- rgamma(n, shape = exp(-param[ , 2]), scale = exp(param[ , 1] + param[ , 2]))

llk_gammals(y = y, param = param, deriv = 4)

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(
  "d0" = function(param){
    sum(llk_gammals(y = y, param = matrix(param, n, 2, byrow = TRUE), deriv = 0)$d0)
  },
  "d1" = function(param){
    colSums(do.call("cbind", llk_gammals(y = y, param = matrix(param, n, 2, byrow = TRUE), deriv = 1)$d1))
  },
  "d2" = function(param){
    colSums(do.call("cbind", llk_gammals(y = y, param = matrix(param, n, 2, byrow = TRUE), deriv = 2)$d2))
  },
  "d3" = function(param){
    colSums(do.call("cbind", llk_gammals(y = y, param = matrix(param, n, 2, byrow = TRUE), deriv = 3)$d3))
  },
  "d4" = function(param){
    colSums(do.call("cbind", llk_gammals(y = y, param = matrix(param, n, 2, byrow = TRUE), deriv = 4)$d4))
  })

check_deriv(obj = obj, param = c(0.5, 0), ord = 1:4)
}
