% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_gev.R
\name{llk_gev}
\alias{llk_gev}
\title{Log-likelihood of the Generalized Extreme Value (GEV) distribution}
\usage{
llk_gev(y, param, deriv = 0, ...)
}
\arguments{
\item{y}{a vector of observations.}

\item{param}{a matrix (or list) with 3 columns (elements), containing the location,
log-scale and shape parameters, in this order.}

\item{deriv}{integer between 0 and 4 indicating the maximum derivative order to
return: 0 only returns \code{d0} (the log-density itself), while 1-4
additionally return \code{d1}-\code{d4}.}
}
\description{
Log-likelihood of the location-scale-shape GEV distribution, and its
derivatives with respect to \code{param = cbind(mu, rho, xi)}, where
\code{mu} is the location, \code{rho = log(sigma)} is the log-scale and
\code{xi} is the shape parameter. The algebra here is ported directly
from the derivative system used internally by \code{mgcv::gevlss}, but
re-expressed as derivatives w.r.t. \code{param} and returned in the list
format used by \code{\link{llk_gaussian}} and friends.
}
\examples{
library(gamFactory)

n <- 10
param <- cbind(mu = rnorm(n), rho = rnorm(n), xi = runif(n, -0.3, 0.3))
y <- param[ , 1] + exp(param[ , 2]) * (((-log(runif(n)))^(-param[ , 3]) - 1) / param[ , 3])

llk_gev(y = y, param = param, deriv = 4)

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(
  "d0" = function(param){
    sum(llk_gev(y = y, param = matrix(param, n, 3, byrow = TRUE), deriv = 0)$d0)
  },
  "d1" = function(param){
    colSums(do.call("cbind", llk_gev(y = y, param = matrix(param, n, 3, byrow = TRUE), deriv = 1)$d1))
  },
  "d2" = function(param){
    colSums(do.call("cbind", llk_gev(y = y, param = matrix(param, n, 3, byrow = TRUE), deriv = 2)$d2))
  },
  "d3" = function(param){
    colSums(do.call("cbind", llk_gev(y = y, param = matrix(param, n, 3, byrow = TRUE), deriv = 3)$d3))
  },
  "d4" = function(param){
    colSums(do.call("cbind", llk_gev(y = y, param = matrix(param, n, 3, byrow = TRUE), deriv = 4)$d4))
  })

check_deriv(obj = obj, param = c(0, 0, 0.1), ord = 1:4)
}
