% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_gumbls.R
\name{llk_gumbls}
\alias{llk_gumbls}
\title{Log-likelihood of the Gumbel location-scale distribution}
\usage{
llk_gumbls(y, param, deriv = 0, ...)
}
\arguments{
\item{y}{a vector of observations.}

\item{param}{a matrix (or list) with 2 columns (elements), containing \code{mu} and
\code{bt}, in this order.}

\item{deriv}{integer between 0 and 4 indicating the maximum derivative order to
return: 0 only returns \code{d0} (the log-density itself), while 1-4
additionally return \code{d1}-\code{d4}.}
}
\description{
Log-likelihood of the Gumbel location-scale model, and its derivatives
with respect to \code{param = cbind(mu, bt)}, where \code{mu} is the mode
and \code{bt = log(beta)} is the log scale of the Gumbel distribution.
Ported directly from the derivative system used internally by
\code{mgcv::gumbls}, but re-expressed as derivatives w.r.t. \code{param}
and returned in the list format used by \code{\link{llk_gaussian}}
and friends. See \code{mgcv::gumbls} for background on the
parametrization.
}
\examples{
library(gamFactory)

n <- 10
param <- cbind(mu = rnorm(n, 2, 0.5), bt = rnorm(n, 0, 0.3))
u <- runif(n)
y <- param[ , 1] - exp(param[ , 2]) * log(-log(u))

llk_gumbls(y = y, param = param, deriv = 4)

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(
  "d0" = function(param){
    sum(llk_gumbls(y = y, param = matrix(param, n, 2, byrow = TRUE), deriv = 0)$d0)
  },
  "d1" = function(param){
    colSums(do.call("cbind", llk_gumbls(y = y, param = matrix(param, n, 2, byrow = TRUE), deriv = 1)$d1))
  },
  "d2" = function(param){
    colSums(do.call("cbind", llk_gumbls(y = y, param = matrix(param, n, 2, byrow = TRUE), deriv = 2)$d2))
  },
  "d3" = function(param){
    colSums(do.call("cbind", llk_gumbls(y = y, param = matrix(param, n, 2, byrow = TRUE), deriv = 3)$d3))
  },
  "d4" = function(param){
    colSums(do.call("cbind", llk_gumbls(y = y, param = matrix(param, n, 2, byrow = TRUE), deriv = 4)$d4))
  })

check_deriv(obj = obj, param = c(2, 0), ord = 1:4)
}
