% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_multinomial.R
\name{llk_multinomial}
\alias{llk_multinomial}
\title{Log-likelihood of a multinomial logistic model}
\usage{
llk_multinomial(y, param, deriv = 0, ...)
}
\arguments{
\item{y}{a vector of observed category labels, integers between 0 (the reference
category) and \code{K = ncol(param)}.}

\item{param}{a matrix (or list) with one column (element) per non-reference category,
containing the corresponding linear predictor (log-odds relative to category 0).}

\item{deriv}{integer between 0 and 4 indicating the maximum derivative order to
return: 0 only returns \code{d0} (the log-density itself), while 1-4
additionally return \code{d1}-\code{d4}.}
}
\description{
Derivatives of the multinomial log-likelihood w.r.t. the linear predictors
(log-odds relative to a reference category 0). Ported from the derivative
computation used internally by \link[mgcv:multinom]{mgcv::multinom}, but re-expressed as
derivatives w.r.t. \code{param} (which plays the role of \code{mu}) rather
than the model coefficients, and returned in the list format used by
\link{llk_gaussian} and friends.
}
\examples{
library(gamFactory)
n <- 10; K <- 3
y <- sample(0:K, n, replace = TRUE)
param <- matrix(rnorm(n * K), n, K)

# Derivatives of multinomial log-lik up to order 4
llk_multinomial(y = y, param = param, deriv = 4)

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(
  "d0" = function(param){
    sum(llk_multinomial(y = y, param = matrix(param, n, K, byrow = TRUE), deriv = 0)$d0)
  },
  "d1" = function(param){
    colSums(do.call("cbind", llk_multinomial(y = y, param = matrix(param, n, K, byrow = TRUE), deriv = 1)$d1))
  },
  "d2" = function(param){
    colSums(do.call("cbind", llk_multinomial(y = y, param = matrix(param, n, K, byrow = TRUE), deriv = 2)$d2))
  },
  "d3" = function(param){
    colSums(do.call("cbind", llk_multinomial(y = y, param = matrix(param, n, K, byrow = TRUE), deriv = 3)$d3))
  },
  "d4" = function(param){
    colSums(do.call("cbind", llk_multinomial(y = y, param = matrix(param, n, K, byrow = TRUE), deriv = 4)$d4))
  })

check_deriv(obj = obj, param = rnorm(K), ord = 1:4)
}
