% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_shash.R
\name{llk_shash}
\alias{llk_shash}
\title{Log-likelihood of the Sinh-Arsinh (shash) distribution}
\usage{
llk_shash(y, param, deriv = 0, phiPen = 0.001)
}
\arguments{
\item{y}{a vector of observations.}

\item{param}{a matrix (or list) with 4 columns (elements), containing \code{mu},
\code{tau}, \code{eps} and \code{phi}, in this order.}

\item{deriv}{integer between 0 and 3 indicating the maximum derivative order to
return: 0 only returns \code{d0} (the log-density itself), while 1-3
additionally return \code{d1}-\code{d3} (4th order derivatives are not
implemented for this family).}

\item{phiPen}{positive coefficient of ridge penalty on phi parameter.}
}
\description{
Log-likelihood of the Sinh-Arsinh (shash) location-scale-skewness-kurtosis
distribution, and its derivatives with respect to
\code{param = cbind(mu, tau, eps, phi)}, where \code{mu} is the location,
\code{tau = log(sigma)} is the log-scale, \code{eps} is the skewness and
\code{phi = log(delta)} is the log-kurtosis. Returned in the list format
used by \code{[gamFactory::llk_gaussian]} and friends.
}
\examples{

library(gamFactory)

##### Simulate data and create object
n <- 10
param <- c(mu = 1, tau = 2, eps = 0.5, phi = 1) # tau = log(sigma), phi = log(delta)
sig <- exp(param["tau"]); del <- exp(param["phi"])
y <- param["mu"] + (del * sig) * sinh((1 / del) * asinh(qnorm(runif(n))) + (param["eps"] / del))

llk_shash(y = y, param = param, deriv = 3)

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(
  "d0" = function(param){
    sum(llk_shash(y = y, param = param, deriv = 0)$d0)
  },
  "d1" = function(param){
    colSums(do.call("cbind", llk_shash(y = y, param = param, deriv = 1)$d1))
  },
  "d2" = function(param){
    colSums(do.call("cbind", llk_shash(y = y, param = param, deriv = 2)$d2))
  },
  "d3" = function(param){
    colSums(do.call("cbind", llk_shash(y = y, param = param, deriv = 3)$d3))
  })

check_deriv(obj = obj, param = param, ord = 1:3)
}
