% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_ztnbinom2.R
\name{llk_ztnbinom2}
\alias{llk_ztnbinom2}
\title{Log-likelihood of a zero-truncated negative binomial model}
\usage{
llk_ztnbinom2(y, param, deriv = 0, ...)
}
\arguments{
\item{y}{XXX}

\item{param}{XXX}

\item{deriv}{XXX}

\item{...}{Can be ignored.}
}
\description{
Many of the derivatives are constructed courtesy of the \code{\link[Deriv:Deriv]{Deriv::Deriv()}} package, with close to zero checks made on this after the second derivative!
}
\examples{
\dontrun{
# Please see the help file for [fam_ztnbinom2()]. 
}

}
