% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_ztpoisson.R
\name{llk_ztpoisson}
\alias{llk_ztpoisson}
\title{Log-likelihood of a zero-truncate Poisson model}
\usage{
llk_ztpoisson(y, param, deriv = 0, ...)
}
\arguments{
\item{param}{XXX.}

\item{deriv}{XXX.}
}
\description{
XXX.
}
\examples{
\dontrun{
# Please see the help file for [fam_ztpoisson()]. 
}

}
