% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikELF.R
\name{logLikELF}
\alias{logLikELF}
\title{Log-likelihood of Extended Log-F (ELF) distribution}
\usage{
logLikELF(y, qu, h)
}
\arguments{
\item{np}{XXX.}
}
\description{
XXX.
}
\examples{
library(gamFactory)
n <- 1000
pars <- c(1, 0.5)
obj <- logLikELF( )$initialize(n, pars, h = 0.7, qu = 0.4)

# Derivatives should match exactly
fdDeriv(obj = derFunWrapper(obj$derObj), 
        param = pars, 
        ord = 1:3)

# Should look fine
der <- derivCheck(np = 100, 
                  parSim = function(n){ cbind(rnorm(n, 0, 1e3), 0.1 + rexp(n, 1)) }, 
                  obj = obj,
                  ord = 1:3, 
                  trans = function(.x){
                    si <- sign(.x)
                    return( si * sqrt(abs(.x)) )
                  }, 
                  qu = 0.99, 
                  h = 0.2, 
                  n = 1000)

par(mfrow = c(2, 2))
for(ii in 1:3) { plot(der[[ii]][ , 1] - der[[ii]][ , 2]) }

par(mfrow = c(2, 2))
for(ii in 1:3) { plot( (der[[ii]][ , 1] - der[[ii]][ , 2]) / 
                         abs( der[[ii]][ , 2] )) }

}
