% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikGAU.R
\name{logLikGAU}
\alias{logLikGAU}
\title{Log-likelihood of the Gaussian distribution}
\usage{
logLikGAU(y = NULL)
}
\arguments{
\item{np}{XXX.}
}
\description{
XXX.
}
\examples{
library(gamFactory)
n <- 1000
pars <- c(1, 0.5)
obj <- logLikGAU( )$initialize(n, pars)

# Derivatives should match exactly
fdDeriv(obj = derFunWrapper(obj$derObj), 
        param = pars, 
        ord = 1:3)

# Should look fine
der <- derivCheck(np = 100, 
                  parSim = function(n){ cbind(rnorm(n, 0, 1e3), 0.1 + rexp(n, 1)) }, 
                  obj = obj,
                  ord = 1:3, 
                  trans = function(.x){
                    si <- sign(.x)
                    return( si * sqrt(abs(.x)) )
                  }, 
                  n = 1000)

par(mfrow = c(2, 2))
for(ii in 1:3) { plot(der[[ii]][ , 1] - der[[ii]][ , 2]) }

par(mfrow = c(2, 2))
for(ii in 1:3) { plot( (der[[ii]][ , 1] - der[[ii]][ , 2]) / 
                         abs( der[[ii]][ , 2] )) }


}
