% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikMARKOV.R
\name{logLikMARKOV}
\alias{logLikMARKOV}
\title{Log-likelihood of a Markov process}
\usage{
logLikMARKOV(y = NULL)
}
\arguments{
\item{np}{XXX.}
}
\description{
XXX.
}
\examples{

library(gamFactory)

# Simulate some data from Markov chain with two states (0, 1)
n <- 5000
x = sort(runif(n, -1, 1))

# Probabilities of staying in state 0 and 1
p1 <- exp(x^2) / (1 + exp(x^2))
p2 <- exp(-x - 0.6*x^2) / (1 + exp(-x - 0.6*x^2))

# Simulate observations 
y0 <- numeric(n)
y0[1] <- 1
for(ii in 2:n){
  if(y0[ii-1] == 0) y0[ii] <- rbinom(1, 1, 1 - p1[ii])
  if(y0[ii-1] == 1) y0[ii] <- rbinom(1, 1, p2[ii])
}

myDat <- data.frame(x = x, y = y0)

# Plot data and probabilities
par(mfrow = c(2, 2))
plot(myDat)
plot(x, p1)
plot(x, p2)

# Create Markov model and fit it
library(mgcViz)
myMRK <- buildGAMLSS( bundleMARKOV )
fit <- gamV(list(y ~ s(x), ~ s(x)), data = myDat, family = myMRK)

# Compare true and estimated probabilities
plot(fit, select = 1, 
     trans = function(.x) exp(coef(fit)[1] + .x) / (1 + exp(coef(fit)[1] + .x))) + 
  l_fitLine() + l_ciLine() + 
  geom_line(data = data.frame(x = x, y = p1), 
            mapping = aes(x=x, y=y), inherit.aes = FALSE, color = "red")

plot(fit, select = 2, 
     trans = function(.x) exp(coef(fit)[11] + .x) / (1 + exp(coef(fit)[11] + .x))) + 
  l_fitLine() + l_ciLine() + 
  geom_line(data = data.frame(x = x, y = p2), 
            mapping = aes(x=x, y=y), inherit.aes = FALSE, color = "red")

}
