% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikSH.R
\name{logLikSH}
\alias{logLikSH}
\title{Log-likelihood of the Sinh-Arsinh (shash) distribution}
\usage{
logLikSH(y = NULL, phiPen = 0.001)
}
\arguments{
\item{y}{a vector of observations.}

\item{phiPen}{positive coefficient of ridge penalty on phi parameter.}
}
\description{
Log-likelihood of the Sinh-Arsinh (shash) distribution
}
\examples{

library(gamFactory)

##### Simulate data and create object
n <- 1000
pars <- c(1, 2, 0.5, 1)
obj <- logLikSH( )$initialize(n, pars)

fdDeriv(obj = derFunWrapper(obj$derObj),
        param = pars,
        ord = 1:3)

### Serious derivatives testing
der <- derivCheck(np = 100,
                  parSim = function(n){
                    tau <- rnorm(n)
                    cbind(rnorm(n, 0, 1e3),
                          tau,
                          exp(tau)*runif(n, -1.5, 1.5),
                          log(runif(n, 0.9, 1.1)))
                  },
                  obj = obj,
                  ord = 1:3,
                  trans = function(.x){
                    si <- sign(.x)
                    return( si * sqrt(abs(.x)) )
                  },
                  n = 100)

par(mfrow = c(2, 2))
for(ii in 1:3) { plot(der[[ii]][ , 1] - der[[ii]][ , 2]) }

par(mfrow = c(2, 2))
for(ii in 1:3) { plot( (der[[ii]][ , 1] - der[[ii]][ , 2]) /
                         abs( der[[ii]][ , 2] )) }

}
