% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLink.R
\name{makeLink}
\alias{makeLink}
\title{Function for setting up link functions}
\usage{
makeLink(link)
}
\arguments{
\item{link}{a character string indicating the link function to be used.}
}
\description{
Extension of \code{mgcv::make.link} including additional links.
}
