% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgks.R
\name{mgks}
\alias{mgks}
\title{Multivariate Gaussian kernel smoother and its derivatives}
\usage{
mgks(y, dist, beta, deriv = 0)
}
\arguments{
\item{y}{A numeric vector of length \code{n0}, or a numeric matrix of dimensions
\code{n} by \code{n0}. Represents the response variable observations. If a matrix,
each row contains a distinct response vector corresponding to that evaluation location.}

\item{dist}{A list of length \code{p} (the number of covariate dimensions). Each
element must be an \code{n} by \code{n0} matrix where the \code{[i, j]}-th entry
is the squared distance between the \code{i}-th evaluation point and the
\code{j}-th reference observation along that specific dimension.}

\item{beta}{A numeric vector of length \code{p} containing the transformation
parameters, where \eqn{\beta_j = \log(1/\sigma_j)} and \eqn{\sigma_j} is the
standard deviation (bandwidth) of the Gaussian kernel for the \code{j}-th dimension.}

\item{deriv}{An integer (0, 1, 2, or 3) indicating the maximum order of
derivatives to compute with respect to \code{beta}. Defaults to 0.}
}
\value{
A list containing the following components:
\itemize{
\item{\code{d0}}{ A numeric vector of length \code{n} containing the kernel smooth evaluations.}
\item{\code{d1}}{ If \code{deriv >= 1}, an \code{n} by \code{p} matrix containing the first-order partial derivatives w.r.t. \code{beta}. Otherwise \code{NULL}.}
\item{\code{d2}}{ If \code{deriv >= 2}, an \code{n} by \code{p*(p+1)/2} matrix containing the second-order derivatives w.r.t. \code{beta} in a packed upper-triangular layout. Otherwise \code{NULL}.}
\item{\code{d3}}{ If \code{deriv >= 3}, an \code{n} by \code{choose(p+2, p-1)} matrix containing the third-order derivatives w.r.t. \code{beta} in a packed configuration. Otherwise \code{NULL}.}
}
}
\description{
Computes a multivariate Gaussian kernel smooth of a response vector
or matrix using pre-calculated coordinate-wise squared distance matrices. It
efficiently provides the smooth values along with analytical derivatives up to
order 3 with respect to the log-inverse bandwidth parameters (\code{beta}).
}
\details{
The function fits a multivariate Gaussian kernel smoother. The weight
\eqn{\alpha_{ij}} assigned to the \code{j}-th reference point when evaluating
at the \code{i}-th location is calculated as follows. Let \eqn{w_k = \exp(2\beta_k) }
be the precision weight for dimension \eqn{k}. The log-kernel evaluation is:
\deqn{\log K_{ij} = -\sum_{k=1}^p w_k \cdot \text{dist}_{k}[i, j]}
\deqn{\alpha_{ij} = \frac{\exp(\log K_{ij} - \max_j \log K_{ij})}{\sum_{q} \exp(\log K_{iq} - \max_j \log K_{ij})}}
The estimated smooth at evaluation row \code{i} is then \eqn{\sum_j \alpha_{ij} y_j}
(or \eqn{\sum_j \alpha_{ij} y_{ij}} if \code{y} is a matrix). Derivatives
are calculated analytically with respect to \code{beta}. Higher-order derivatives
(\code{d2} and \code{d3}) are stored in a packed, symmetric storage format to
minimize memory usage.
}
\examples{
### Example
# Simulate values of 2D function
n <- 1000
X0 <- cbind(runif(n, -1, 1), runif(n, -4, 4))
trueF <- function(x) 3 * x[ , 1] + x[ , 2]^2
y <- trueF(X0)

# Evaluate kernel smooth on a grid
ngr <- 40
xseq1 <- seq(-1, 1, length.out = ngr)
xseq2 <- seq(-4, 4, length.out = ngr)
X <- as.matrix(expand.grid(xseq1, xseq2))

# Here beta[j] is log(1 / sigma[j]) where sigma is standard deviation
beta <- c(log(20), log(5))
dist <- lapply(1:ncol(X), function(dd) 
                t(sapply(1:nrow(X), function(ii) (X[ii, dd] - X0[ , dd])^2 )))
fit <- mgks(y = y, dist = dist, beta = beta, deriv = 3)

par(mfrow = c(2, 1))
image(x = xseq1, y = xseq2, z = matrix(fit$d0, ngr, ngr), main = "Estimated")
image(x = xseq1, y = xseq2, z = matrix(trueF(X), ngr, ngr), main = "Truth")

### Checking derivatives by finite differences

# Simulate values of 3D function
n <- 1000
X0 <- cbind(runif(n, -1, 1), runif(n, -4, 4), runif(n, -2, 2))
trueF <- function(x) 3 * x[ , 1] + x[ , 2]^2 - 0.5 * x[ , 3]^2
y <- trueF(X0)
# Define grid over which derivative will be evaluated
ngr <- 5
xseq1 <- seq(-1, 1, length.out = ngr)
xseq2 <- seq(-4, 4, length.out = ngr)
xseq3 <- seq(-2, 2, length.out = ngr)
X <- as.matrix(expand.grid(xseq1, xseq2, xseq3))

dist <- lapply(1:ncol(X), function(dd) 
                t(sapply(1:nrow(X), function(ii) (X[ii, dd] - X0[ , dd])^2 )))

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(d0 = function(param){
   sum(mgks(y = y, dist = dist, beta = param, deriv = 0)$d0)
},
d1 = function(param){
   tmp <- colSums(mgks(y = y, dist = dist, beta = param, deriv = 1)$d1)
   as.list(tmp)
},
d2 = function(param){
   tmp <- colSums(mgks(y = y, dist = dist, beta = param, deriv = 2)$d2)
   as.list(tmp)
},
d3 = function(param){
   tmp <- colSums(mgks(y = y, dist = dist, beta = param, deriv = 3)$d3)
   as.list(tmp)
})

# Check derivatives up to order 3
beta <- c(log(10), log(5), log(5))
(tmp <- check_deriv(obj = obj, param = beta, ord = 1:3))

par(mfrow = c(2, 2))
for(ii in 1:3){
  plot(tmp[[ii]][ , 2], tmp[[ii]][ , 1] - tmp[[ii]][ , 2],
       main = paste0("mgks deriv ", ii), ylab = "Error",
       xlab = "Deriv value")
  abline(h = 0)
}

# Test case where we have a different y vector for each location
# at which we want to evaluate the smooth

y_mat <- do.call("rbind", lapply(1:nrow(X), function(nouse) y))
y_mat <- y_mat + rnorm(ncol(y_mat) * nrow(y_mat))

# Check derivatives up to order 3
beta <- c(log(10), log(5), log(5))
(tmp <- check_deriv(obj = obj, param = beta, ord = 1:3))

par(mfrow = c(2, 2))
for(ii in 1:3){
  plot(tmp[[ii]][ , 2], tmp[[ii]][ , 1] - tmp[[ii]][ , 2],
       main = paste0("mgks deriv ", ii), ylab = "Error",
       xlab = "Deriv value")
  abline(h = 0)
}
}
