% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_var.R
\name{pen_var}
\alias{pen_var}
\alias{pen_var_outer}
\title{Penalty on variance of the inner linear predictor}
\usage{
pen_var(o, v, deriv = 0)

pen_var_outer(o, v, DaDr)
}
\arguments{
\item{o}{the effect object.}

\item{v}{target variance of the inner linear predictor.}

\item{deriv}{number of derivatives to be calculated.}

\item{DaDr}{derivative for parameters of the inner linear predictor
w.r.t. the smoothing parameter.}
}
\description{
Penalty on variance of a nested effect.
}
