% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_vargen.R
\name{pen_vargen}
\alias{pen_vargen}
\alias{pen_vargen_outer}
\title{Penalty on variance of a non-linear effect}
\usage{
pen_vargen(o, v, deriv = 0)

pen_vargen_outer(o, v, DaDr)
}
\arguments{
\item{v}{target variance of the projected covariates.}

\item{deriv}{number of derivatives to be calculated.}

\item{DaDr}{derivative for single index vector w.r.t. smoothing parameter.}

\item{a}{single index parameter vector of dimension p.}

\item{x}{an n by p matrix of covariates to be projected.}
}
\description{
Penalty on variance of a single index vector.
}
