% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_varSI.R
\name{pen_varSI}
\alias{pen_varSI}
\alias{pen_varSI_outer}
\title{Penalty on variance of a single index vector}
\usage{
pen_varSI(a, x, v, deriv = 0)

pen_varSI_outer(a, x, DaDr)
}
\arguments{
\item{a}{single index parameter vector of dimension p.}

\item{x}{an n by p matrix of covariates to be projected.}

\item{v}{target variance of the projected covariates.}

\item{deriv}{number of derivatives to be calculated.}

\item{DaDr}{derivative for single index vector w.r.t. smoothing parameter.}
}
\description{
Penalty on variance of a single index vector.
}
