% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_gam_nl.R
\name{postproc_gam_nl}
\alias{postproc_gam_nl}
\title{Post-process a GAM containing non-standard effects}
\usage{
postproc_gam_nl(o, info)
}
\description{
Post-process a GAM containing non-standard effects
}
