% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_all_times.R
\name{pre_all_times}
\alias{pre_all_times}
\title{Predict fitted values at all time points from a nested GAM-NL model}
\usage{
pre_all_times(object, X_ori, w_ori)
}
\arguments{
\item{object}{A fitted \code{gamnl} model object obtained from
\code{\link[gamFactory]{gamnl}} or similar fitting functions.}

\item{X_ori}{A matrix of covariates corresponding to the inner
linear transformation component.}

\item{w_ori}{A matrix of covariates corresponding to the outer
exponential smoothing component.}
}
\value{
A numeric vector containing the predicted values for all time points
in the same order as the input rows of \code{X_ori} and \code{w_ori}.
}
\description{
smoothing components.
}
