% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gamnl.R
\name{predict.gamnl}
\alias{predict.gamnl}
\title{Predictions from GAM models with non-linear effects}
\usage{
\method{predict}{gamnl}(object, newdata, type = "link", se.fit = FALSE, ...)
}
\arguments{
\item{...}{further arguments to be passed to \link[mgcv:gam]{mgcv::gam}.}

\item{formula, family, data}{same arguments as in \link[mgcv:gam]{mgcv::gam}.}
}
\description{
Predictions from GAM models with non-linear effects
}
