% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_matrix_mgks_smooth.R
\name{Predict.matrix.mgks.smooth}
\alias{Predict.matrix.mgks.smooth}
\title{Predict using MGKS smooth effects}
\usage{
\method{Predict.matrix}{mgks.smooth}(object, data)
}
\description{
Predict using MGKS smooth effects
}
