% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_matrix_nexpsm_smooth.R
\name{Predict.matrix.nexpsm.smooth}
\alias{Predict.matrix.nexpsm.smooth}
\title{Predict using nested exponential smoothing effects}
\usage{
\method{Predict.matrix}{nexpsm.smooth}(object, data)
}
\description{
Predict using nested exponential smoothing effects
}
