% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_matrix_si_smooth.R
\name{Predict.matrix.si.smooth}
\alias{Predict.matrix.si.smooth}
\title{Predict using single index effects}
\usage{
\method{Predict.matrix}{si.smooth}(object, data)
}
\description{
Predict using single index effects
}
