% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_info.R
\name{prep_info}
\alias{prep_info}
\title{Get information to build family for GAM with non-standard effect}
\usage{
prep_info(o)
}
\description{
Get information to build family for GAM with non-standard effect
}
