% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_link.R
\name{prepare_link}
\alias{prepare_link}
\title{Preparing link function(s) of a response family}
\usage{
prepare_link(link, oklinks, np, nam)
}
\arguments{
\item{link}{A character vector of link function(s) to be used for the response family.}

\item{oklinks}{A list of character vectors, each containing the allowed link functions.}

\item{np}{Number of parameters of the family.}

\item{nam}{Name of the family.}
}
\description{
Internal function used to check whether a link function is within
the set of allowed link functions for a given response family. It called
\code{fix_family_link} to ensure that the link function has the correct derivatives.
}
\examples{
library(gamFactory)
prepare_link(link = "logitab(1.2, 4)", "logitab(3, 5)", 1, "gaussian")
prepare_link(link = "loginva(1.2)", "loginva", 1, "gaussian")

}
