% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_construct_inter_dlinear_smooth_spec.R
\name{smooth.construct.inter_dlinear.smooth.spec}
\alias{smooth.construct.inter_dlinear.smooth.spec}
\title{Nested single-index in tensor product}
\usage{
\method{smooth.construct}{inter_dlinear.smooth.spec}(object, data, knots)
}
\description{
Nested single-index in tensor product
}
