% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_construct_inter_mgks_smooth_spec.R
\name{smooth.construct.inter_mgks.smooth.spec}
\alias{smooth.construct.inter_mgks.smooth.spec}
\title{Nested interactive mgks smoothing effect}
\usage{
\method{smooth.construct}{inter_mgks.smooth.spec}(object, data, knots)
}
\description{
Nested interactive mgks smoothing effect
}
