% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_construct_mgks_smooth_spec.R
\name{smooth.construct.mgks.smooth.spec}
\alias{smooth.construct.mgks.smooth.spec}
\title{Nested MGKS effects for mgcv}
\usage{
\method{smooth.construct}{mgks.smooth.spec}(object, data, knots)
}
\description{
Nested MGKS effects for mgcv
}
\examples{
#n <- 100
#n0 <- 50
#p <- 5
#dat <- data.frame(y = rnorm(n))
#dat$Xi <- matrix(rnorm(n*p), n, p)
#X0 <- matrix(rnorm(n0*p), n0, p)
#x <- rnorm(n0)
#aaa <- smoothCon(s(Xi, bs = "mgks", m = c(3, 2),
#            xt = list(si = list(X0 = X0, x = x), sumConv = FALSE)),
#          data = dat)
}
