% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackFamily.R
\name{stackFamily}
\alias{stackFamily}
\title{Creates a family for stack effects}
\usage{
stackFamily(X, familyDeriv, link = "identity")
}
\arguments{
\item{X}{Model Matrix of the nonnegative linear combination}

\item{familyDeriv}{Families of the distribution of the response variable
with respect to its parameters. At the moment only Gaussian and SHASH families
are possible. Customed families have been written through createGaussian and createSHASH}

\item{link}{At the moment only the identity link is possible}
}
\value{
A family to give as family argument to mgcv::gam
}
\description{
Creates a family to be used in mgcv::gam to estimate
a generalized additive model with stack effects.
}
