% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackFamilyGaussian.R
\name{stackFamilyGaussian}
\alias{stackFamilyGaussian}
\title{Creates stack effects for a Gaussian family}
\usage{
stackFamilyGaussian(X, link = "identity")
}
\arguments{
\item{X}{Model Matrix of the nonnegative linear combination}

\item{link}{At the moment only the identity link is possible}
}
\value{
A family to give as family argument to mgcv::gam
}
\description{
Creates a family to be used in mgcv::gam to estimate a generalized additive model
with response variable having Gaussian distribution and stack nonnegative effects.
}
