% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackPredictiveFamily.R
\name{stackPredictiveFamily}
\alias{stackPredictiveFamily}
\title{Creates a family for stacking predictive distributions}
\usage{
stackPredictiveFamily(logP, rho = NULL)
}
\arguments{
\item{logP}{n x K matrix of the log-predictive distributions of the single experts.
n is the total number of available observations, while K is the number of experts}
}
\value{
A family to give as family argument to mgcv::gam
}
\description{
Creates a family to be used in mgcv::gam
for stacking predictive distributions.
}
