% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmat2vec.R
\name{symmat2vec}
\alias{symmat2vec}
\title{Transform symmetric matrix to vector}
\usage{
symmat2vec(H)
}
\arguments{
\item{H}{a symmetric matrix.}
}
\value{
A vector of unique elements of H, by row.
}
\description{
The input is a symmetric matrix, the output is a vector of its unique elements by row.
}
\examples{
1:3 \%*\% t(1:3)
symmat2vec(1:3 \%*\% t(1:3))

}
