% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGeneralFamilyDeriv.R
\name{testGeneralFamilyDeriv}
\alias{testGeneralFamilyDeriv}
\title{Test derivatives of a general family to be used in mgcv}
\usage{
testGeneralFamilyDeriv(
  fam,
  nsim = 30,
  ndata = 100,
  p = 2,
  seed = 0,
  nlp = 3,
  ntheta = NULL
)
}
\arguments{
\item{fam}{A general family}

\item{nsim}{Number of parameters to simulate
(both for regression coefficients and smoothing parameters)
at which calculate derivatives}

\item{ndata}{Number of data to generate}

\item{p}{Number of regression coefficients for each linear predictor}

\item{seed}{Set a seed for reproducibility}

\item{ntheta}{Number of additional parameters of the family (to be implemented)}
}
\value{
A comparison between derivatives calculated according to the ll function of the family
and numerical derivatives calculated using jacobian function
}
\description{
Test derivatives of a general family to be used in mgcv
}
