% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_der_fun.R
\name{createGPDs}
\alias{createGPDs}
\alias{wrap_der_fun}
\title{Wrapper to fdDeriv()}
\usage{
wrap_der_fun(obj, dropH = FALSE)
}
\arguments{
\item{obj}{for instance the output of createGPD.}

\item{dropH}{if TRUE the Hessian will be transformed to a vector using flattenSymMat.}
}
\description{
This is a wrapper useful for checking derivative of families with the fdDeriv() function.
}
