% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapDerFun.R
\name{createGPDs}
\alias{createGPDs}
\alias{wrapDerFun}
\title{Wrapper to fdDeriv()}
\usage{
wrapDerFun(obj, dropHessian = FALSE)
}
\arguments{
\item{obj}{for instance the output of createGPD.}

\item{dropHessian}{if TRUE the Hessian will be transformed to a vector using flattenSymMat.}
}
\description{
This is a wrapper useful for checking derivative of families with the fdDeriv() function.
}
