% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_bspline.R
\name{basis_bspline}
\alias{basis_bspline}
\title{Get B-spline design matrix constructor}
\usage{
basis_bspline(knots, m)
}
\arguments{
\item{knots}{positions of all the knots, both inner and outer.}

\item{m}{order of the B-spline basis. Here \code{m = 0} for piecewise linear,
\code{m = 1} for quadratic and so on. Number of continuous derivatives
is \code{m}, e.g., if \code{m = 4} fourth order derivatives are continuous
but fifth are not.}
}
\value{
A function that take arguments \code{x} and \code{deriv} and returns
a B-splines model matrix and its first \code{deriv} derivatives w.r.t. \code{x}.
}
\description{
Get B-spline design matrix constructor
}
\examples{
library(gamFactory)

xseq <- seq(-6, 6, 0.01)
k <- 10             
m <- 4
ko <- k + m + 2
knots <- seq(xseq[1], tail(xseq,1), length.out = ko)
tmp <- basis_bspline(knots = knots, m = m)$evalX(x = xseq, deriv = 4)

par(mfrow = c(3, 2)) 
matplot(xseq, tmp$X0, type = 'l') # Design matrix
matplot(xseq, tmp$X1, type = 'l') # 1st derivative w.r.t. x 
matplot(xseq, tmp$X2, type = 'l') # 2nd derivative w.r.t. x 
matplot(xseq, tmp$X3, type = 'l') # 3rd derivative w.r.t. x
matplot(xseq, tmp$X4, type = 'l') # 4rd derivative w.r.t. x
rug(knots)

# Checking if derivatives are right by finite differences: it seems so
l0 <- basis_bspline(knots = knots, m = m)$evalX(x = xseq, deriv = 4)
lm <- basis_bspline(knots = knots, m = m)$evalX(x = xseq-1e-5, deriv = 4)
lp <- basis_bspline(knots = knots, m = m)$evalX(x = xseq+1e-5, deriv = 4)
par(mfrow = c(2, 2))
plot(l0$X1, (lp$X0 - lm$X0) / 2e-5)
abline(0, 1, col = 2)
plot(l0$X2, (lp$X1 - lm$X1) / 2e-5)
abline(0, 1, col = 2)
plot(l0$X3, (lp$X2 - lm$X2) / 2e-5)
abline(0, 1, col = 2)
plot(l0$X4, (lp$X3 - lm$X3) / 2e-5)
abline(0, 1, col = 2)

}
