% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_family_nl.R
\name{build_family_nl}
\alias{build_family_nl}
\title{Function for building GAM families containing non-standard effects}
\usage{
build_family_nl(y_true = NULL, bundle, info, lamVar = 1e+05, lamRidge = 1e-05)
}
\description{
Function for building GAM families containing non-standard effects
}
