% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_gpd.R
\name{bundle_gpd}
\alias{bundle_gpd}
\title{Ingredients for Generalized Pareto family}
\usage{
bundle_gpd()
}
\description{
XXX.
}
\examples{
library(gamFactory)

# Simulate some data
n <- 2000
myDat <- data.frame(x = runif(n, -1, 1))
myDat$y <-  rexp(n, 1 / (0.3 + abs(myDat$x)^2))
plot(myDat)

# Create GPD family and fit it
fit <- gam_nl(list(y ~ s(x), ~ s(x)), data = myDat, family = build_family(bundle_gpd()))
plot(fit, pages = 1)

# Compare true (black) and estimated (red) quantiles
qu <- 0.95
xseq <- seq(-1, 1, length.out = 1000)
plot(myDat, col = "grey")
lines(xseq, qexp(qu, 1 / (0.3 + abs(xseq)^2)), col = 2)
lines(xseq, fit$family$qf(qu,
                          mu = predict(fit, newdata = data.frame(x = xseq),
                                       type = "response")),
      col = 3)

}
