% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_deriv.R
\name{check_deriv}
\alias{check_deriv}
\title{Comparing analytic with numerical finite differences}
\usage{
check_deriv(obj, param, ord = 1)
}
\arguments{
\item{obj}{the function to be differentiated.}

\item{param}{vector of parameters at which we should calculate the derivatives.}

\item{ord}{the order to derivatives to be checked.}
}
\description{
Function for getting higher derivatives by finite differences.
}
