% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicago_data.R
\name{chicago_data}
\alias{chicago_data}
\title{Process chicago data from gamair package}
\usage{
chicago_data()
}
\description{
Process chicago data from gamair package
}
\examples{
dat <- chicago_data()
dim(dat)
str(dat)
}
