% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/der_singleIndex (casa's conflicted copy
%   2020-07-10).R, R/der_singleIndex.R
\name{der.singleIndex}
\alias{der.singleIndex}
\title{Derivatives of single index effects}
\usage{
\method{der}{singleIndex}(o, llk, deriv = 1, param = NULL)

\method{der}{singleIndex}(o, llk, deriv = 1, param = NULL)
}
\description{
Derivatives of single index effects

Derivatives of single index effects
}
