% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_inter_dlinear.R
\name{DHessDrho.inter_dlinear}
\alias{DHessDrho.inter_dlinear}
\title{❗️Maybe we can remove this function if we return "si" class in the end of "eff_inter"}
\usage{
\method{DHessDrho}{inter_dlinear}(o, llk, DbDr)
}
\description{
Derivative of inter_dlinear Hessian w.r.t smoothing parameters
}
