% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{DHessDrho}
\alias{DHessDrho}
\title{Derivatives of Hessian w.r.t. smoothing parameters}
\usage{
DHessDrho(o, ...)
}
\arguments{
\item{o}{object we want to get the derivatives of.}

\item{...}{arguments to be passed to methods.}
}
\description{
Derivatives of Hessian w.r.t. smoothing parameters
}
