% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_nexpsm.R
\name{eff_nexpsm}
\alias{eff_nexpsm}
\title{Build nested adaptive exponential smooth effect}
\usage{
eff_nexpsm(y, Xi, basis, x0 = NULL, times = NULL)
}
\description{
Build nested adaptive exponential smooth effect
}
