% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_si_nexp.R
\name{eff_si_nexp}
\alias{eff_si_nexp}
\title{Build nested adaptive linear transform and exponential smooth effect}
\usage{
eff_si_nexp(
  X_si,
  X_nexp,
  basis,
  Z0,
  times = NULL,
  alpha_center = NULL,
  positive_si = FALSE
)
}
\description{
Build nested adaptive linear transform and exponential smooth effect
}
