% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_stand.R
\name{eff_stand}
\alias{eff_stand}
\title{Build standard effects}
\usage{
eff_stand(X)
}
\arguments{
\item{X}{model matrix, such that effect will be \code{f = X \%*\% beta}.}
}
\description{
Build standard effects
}
