% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_family_link.R
\name{fix_family_link}
\alias{fix_family_link}
\title{Preparing response family function}
\usage{
fix_family_link(fam)
}
\arguments{
\item{fam}{XXX.}
}
\description{
Extension of \code{mgcv::fix.family.link} including additional links.
}
