% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_nl.R
\name{gam_nl}
\alias{gam_nl}
\title{Fit GAM model with non-linear effects}
\usage{
gam_nl(formula, family = gaulss(), data = list(), ...)
}
\arguments{
\item{formula, family, data}{same arguments as in \link[mgcv:gam]{mgcv::gam}.}

\item{...}{further arguments to be passed to \link[mgcv:gam]{mgcv::gam}.}
}
\description{
Fit GAM model with non-linear effects
}
