% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_binomial.R
\name{llk_binomial}
\alias{llk_binomial}
\title{Log-likelhood of a Gaussian model}
\usage{
llk_binomial(y, param, n, deriv = 0, ...)
}
\arguments{
\item{param}{XXX.}

\item{deriv}{XXX.}
}
\description{
XXX.
}
\examples{
library(gamFactory)
nobs <- 10
n <- 50
y <- rbinom(nobs, size = n, prob = 0.3)
param <- c(0.4)

# Derivatives of Gaussian log-lik up to order 3
llk_binomial(y = y, param = param, n = n, deriv = 3)

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(
  "d0" = function(param){
    sum(llk_binomial(y = y, param = param, n = n, deriv = 0)$d0)
  },
  "d1" = function(param){
    colSums(do.call("cbind", llk_binomial(y = y, param = param, n = n, deriv = 1)$d1))
    
  },
  "d2" = function(param){
    colSums(do.call("cbind", llk_binomial(y = y, param = param, n = n, deriv = 2)$d2))
    
  },
  "d3" = function(param){
    colSums(do.call("cbind", llk_binomial(y = y, param = param, n = n, deriv = 3)$d3))
  })

check_deriv(obj = obj, param = param, ord = 1:3)

}
