% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_gpd.R
\name{llk_gpd}
\alias{llk_gpd}
\title{Log-likelihood of Generalized Pareto Distribution}
\usage{
llk_gpd(y, param, deriv = 0, ...)
}
\arguments{
\item{np}{XXX.}
}
\description{
XXX.
}
\examples{
library(gamFactory)
n <- 5
param <- c(1.5, 0.3)
y <- rexp(n)

# Derivatives of Gaussian log-lik up to order 3
llk_gpd(y = y, param = param, deriv = 3)

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(
  "d0" = function(param){
    sum(llk_gpd(y = y, param = param, deriv = 0)$d0)
  },
  "d1" = function(param){
    colSums(do.call("cbind", llk_gpd(y = y, param = param, deriv = 1)$d1))

  },
  "d2" = function(param){
    colSums(do.call("cbind", llk_gpd(y = y, param = param, deriv = 2)$d2))

  },
  "d3" = function(param){
    colSums(do.call("cbind", llk_gpd(y = y, param = param, deriv = 3)$d3))
  })

check_deriv(obj = obj, param = param, ord = 1:3)

}
