% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_poisson.R
\name{llk_poisson}
\alias{llk_poisson}
\title{Log-likelhood of a Gaussian model}
\usage{
llk_poisson(y, param, deriv = 0, ...)
}
\arguments{
\item{param}{XXX.}

\item{deriv}{XXX.}
}
\description{
XXX.
}
\examples{
library(gamFactory)
n <- 10
y <- rpois(n, 15)
param <- c(12.5) # mu and 1/sigma

# Derivatives of Gaussian log-lik up to order 3
llk_poisson(y = y, param = param, deriv = 3)

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(
  "d0" = function(param){
    sum(llk_poisson(y = y, param = param, deriv = 0)$d0)
  },
  "d1" = function(param){
    colSums(do.call("cbind", llk_poisson(y = y, param = param, deriv = 1)$d1))
    
  },
  "d2" = function(param){
    colSums(do.call("cbind", llk_poisson(y = y, param = param, deriv = 2)$d2))
    
  },
  "d3" = function(param){
    colSums(do.call("cbind", llk_poisson(y = y, param = param, deriv = 3)$d3))
  })

check_deriv(obj = obj, param = param, ord = 1:3)


}
