% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_shash.R
\name{llk_shash}
\alias{llk_shash}
\title{Log-likelihood of the Sinh-Arsinh (shash) distribution}
\usage{
llk_shash(y, param, deriv = 0, phiPen = 0.001)
}
\arguments{
\item{y}{a vector of observations.}

\item{phiPen}{positive coefficient of ridge penalty on phi parameter.}
}
\description{
Log-likelihood of the Sinh-Arsinh (shash) distribution
}
\examples{

library(gamFactory)

##### Simulate data and create object
n <- 10
param <- c(1, exp(2), 0.5, exp(1))
y <- param[1] + (param[4] * param[2]) * sinh((1/param[4]) * asinh(qnorm(runif(n))) + (param[3]/param[4]))

llk_shash(y = y, param = param, deriv = 3)

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(
  "d0" = function(param){
    sum(llk_shash(y = y, param = param, deriv = 0)$d0)
  },
  "d1" = function(param){
    colSums(do.call("cbind", llk_shash(y = y, param = param, deriv = 1)$d1))
  },
  "d2" = function(param){
    colSums(do.call("cbind", llk_shash(y = y, param = param, deriv = 2)$d2))
  },
  "d3" = function(param){
    colSums(do.call("cbind", llk_shash(y = y, param = param, deriv = 3)$d3))
  })

check_deriv(obj = obj, param = param, ord = 1:3)
}
