% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_ztpoisson.R
\name{llk_ztpoisson}
\alias{llk_ztpoisson}
\title{Log-likelihood of a zero-truncated Poisson model}
\usage{
llk_ztpoisson(y, param, deriv = 0, ...)
}
\arguments{
\item{y}{XXX}

\item{param}{XXX}

\item{deriv}{XXX}

\item{...}{Can be ignored.}
}
\description{
XXX
}
\examples{
\dontrun{
# Please see the help file for [fam_ztpoisson()]. 
}

}
