% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_link.R
\name{make_link}
\alias{make_link}
\title{Function for setting up link functions}
\usage{
make_link(link)
}
\arguments{
\item{link}{a character string indicating the link function to be used.}
}
\description{
Extension of \code{mgcv::make.link} including additional links.
}
