% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gamnl.R
\name{predict.gamnl}
\alias{predict.gamnl}
\title{Predictions from GAM models with non-linear effects}
\usage{
\method{predict}{gamnl}(object, newdata, type = "link", se.fit = FALSE, ...)
}
\arguments{
\item{object}{XXX}

\item{newdata}{XXX}

\item{type}{XXX}

\item{se.fit}{XXX}

\item{...}{Can be ignored}
}
\description{
Predictions from GAM models with non-linear effects
}
