% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_link.R
\name{prepare_link}
\alias{prepare_link}
\title{Preparing link function(s) of a response family}
\usage{
prepare_link(link, oklinks, np, nam)
}
\arguments{
\item{np}{XXX.}
}
\description{
XXX.
}
\examples{
library(gamFactory)
prepare_link(link = "logitab(1.2, 4)", "logitab(3, 5)", 1, "gaussian")
prepare_link(link = "loginva(1.2)", "loginva", 1, "gaussian")

}
