% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_nest.R
\name{s_nest}
\alias{s_nest}
\title{Defining nested smooths in GAM formulae}
\usage{
s_nest(..., trans, k, m, sarg)
}
\arguments{
\item{...}{a list of variables that are used to defined the nested smooth effect.}

\item{trans}{the type of covariate transformation to use (see, e.g., \link{trans_mgks}).}

\item{k}{the number of basis functions used to construct the outer smooth effect.}

\item{m}{a vector of two positive integers. The first indicates the order of the
outer B-spline basis (see \link{basis_bspline}), the second is the order of the corresponding P-spline
penalty.}

\item{sarg}{a list of further arguments to be passed to the effect constructor.}
}
\description{
Function used to defined nested smooth effects. It works similarly to
\link{mgcv::s}.
}
