% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_construct_inter_nexp_smooth_spec.R
\name{smooth.construct.inter_nexp.smooth.spec}
\alias{smooth.construct.inter_nexp.smooth.spec}
\title{Nested interactive adaptive exponential smoothing effect}
\usage{
\method{smooth.construct}{inter_nexp.smooth.spec}(object, data, knots)
}
\description{
Nested interactive adaptive exponential smoothing effect
}
