% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_construct_nexpsm_smooth_spec.R
\name{smooth.construct.nexpsm.smooth.spec}
\alias{smooth.construct.nexpsm.smooth.spec}
\title{Nested adaptive exponential smoothing effects for mgcv}
\usage{
\method{smooth.construct}{nexpsm.smooth.spec}(object, data, knots)
}
\description{
Nested adaptive exponential smoothing effects for mgcv
}
