% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_construct_si_smooth_spec.R
\name{smooth.construct.si.smooth.spec}
\alias{smooth.construct.si.smooth.spec}
\title{Single index effects for mgcv}
\usage{
\method{smooth.construct}{si.smooth.spec}(object, data, knots)
}
\description{
Single index effects for mgcv
}
