% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_xxx.R
\name{trans_linear}
\alias{trans_linear}
\alias{trans_mgks}
\alias{trans_nexpsm}
\alias{trans_linear_nexpsm}
\title{Specifying inner transformations}
\usage{
trans_linear(pord, S, alpha, a0)

trans_mgks(y0, alpha)

trans_nexpsm(S, alpha)

trans_linear_nexpsm(
  pord = NULL,
  S_si = NULL,
  S_nexp = NULL,
  alpha_nexp = NULL,
  alpha_si = NULL,
  alpha_scale = NULL,
  center = FALSE,
  alpha_center = NULL,
  n_si = NULL,
  n_nexp = NULL,
  Z0 = NULL,
  positive_si = FALSE
)
}
\arguments{
\item{pord}{The order of the differences of the linear transformation parameters
that will be penalised. E.g., \code{pord = 1} corresponds to penalising the squared
first order differences \eqn{(\alpha_k - \alpha_{k-1})^2}. Ignored if argument \code{S}
is supplied.}

\item{S}{Positive semi-definite matrix used to specify a generalised ridge
penalty on the transformation parameters. In \code{trans_linear}, if \code{S} is not supplied, the \code{pord}-order differences
between parameters are penalised, see the \code{pord} argument.}

\item{alpha}{Vector containing the initial values for the parameters of the transformation.}

\item{y0}{Vector of \eqn{n} observations corresponding to the rows of the variables in \code{s_nest}.}

\item{S_si}{Penalty matrix used in trans_linear_nexpsm. For the linear transformation parameters.}

\item{S_nexp}{Penalty matrix used in expsmooth. For the linear transformation parameters.}

\item{alpha_nexp}{Initial values used in trans_linear_nexpsm. For expsmooth parameters.}

\item{alpha_si}{Initial values used in trans_linear_nexpsm. For linear transformation parameters.}

\item{n_si}{Length of \code{alpha_si} if \code{alpha_si} is not provided.}

\item{n_nexp}{Length of \code{alpha_nexp} if \code{alpha_nexp} is not provided.}
}
\description{
Functions to specify nested covariate transformations, to be used
within \link{s_nest}.
}
\details{
The types of transformations currently provided are four: the linear transformation,
the multivariate kernel smooth transformation, the exponential smoothing transformation, and a double
nested effect which combines a single index and exponential smooth. Specifically:
\itemize{
\item{\code{trans_linear}}{ implements a linear transformation \eqn{X^\top \alpha}, the simplest type of
transformation which can be used to specify a single index vector (i.e. a projection) or a linear effect.}
\item{\code{trans_mgks}}{ implements a multivariate kernel smooth transformation based on the response variable
observation vector \code{y0} and corresponding distance matrix. The transformation is defined as
\deqn{\tilde{s}(\mathbf{x}_i) = \frac{\sum_{j\in \mathcal{N}_i}K_{\mathbf{a}}(\mathbf{x}_{i},
        \mathbf{x}_{j})y0_{j}}{\sum_{q\in \mathcal{N}_i}K_{\mathbf{a}}(\mathbf{x}_{i},\mathbf{x}_{q})},}
with \eqn{K_{\mathbf{a}}} denoting the kernel of a multivariate p.d.f. parameterised by
\eqn{\mathbf{a}}, and \eqn{\mathcal{N}_i} the index set of neighbours of \eqn{\mathbf{x}_i}.}
\item{\code{trans_nexpsm}}{ implements an exponential smoothing transformation, defined as
\deqn{\tilde{s}(x_i) = \tilde{s}_i = \omega_{i}\tilde{s}_{i-1}+(1-\omega_{i})x_{i}, \qquad \text{for } i \geq 1,}
with \eqn{\tilde{s}_0 = x_0} and \eqn{\omega_i \in (0, 1)}. The smoothing factor can be defined as
\eqn{\omega_i = \phi(\tilde{\mathbf{x}}_i^\top \mathbf{a})}, where \eqn{\phi} is the logistic function,
and \eqn{\tilde{\mathbf{x}}_i} is a given vector.}
\item{\code{trans_linear_nexpsm}}{ implements a combination of a linear transformation and an exponential
smoothing transformation. The transformation is defined as
\deqn{\tilde{s}({\bf x}_i, {\bf w}_i) = \tilde{s}_i = \omega_i \tilde{s}_{i-1} + (1-\omega_i)z_i,}
where \eqn{z_i = {\bf x}_i^\top \bm a_{\mathrm{si}}}, for \eqn{i \geq 1,\ \tilde{s}_0 = z_0},
and the smoothing factor is defined as \eqn{\omega_i = \phi({\bf w}_i^\top \bm a_{\mathrm{nexp}}),\ \omega_i \in (0,1)}.
As in the previous case, the smoothing factor can be defined as
\eqn{\omega_i = \phi(\tilde{\mathbf{x}}_i^\top {\mathbf{a}_{\mathrm{si}})}}, where \eqn{\phi} is the logistic function
and \eqn{\tilde{\mathbf{x}}_i} is a given vector.}
}

\code{trans_nexpsm} and \code{trans_linear_nexpsm} allow for different time granularities in the inner exponential and outer smooths.
It is particularly useful for handling data structures where the response variable is observed at a
different frequency than the covariates driving the smoothing rate (e.g., daily response with hourly covariates).
}
