% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_xxx.R
\name{trans_mgks}
\alias{trans_mgks}
\alias{trans_nexpsm}
\alias{trans_linear}
\title{Specifying a MGKS transformation}
\usage{
trans_mgks(n0, vr, init)

trans_nexpsm(S, vr, init)

trans_linear(pord, S, vr, init)
}
\description{
Specifying a MGKS transformation
}
