% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_xxx.R
\name{trans_mgks}
\alias{trans_mgks}
\alias{trans_nexpsm}
\alias{trans_linear}
\title{Specifying a MGKS transformation}
\usage{
trans_mgks(X0, y0, alpha)

trans_nexpsm(S, alpha)

trans_linear(pord, S, alpha)
}
\description{
Specifying a MGKS transformation
}
