% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_xxx.R
\name{trans_linear}
\alias{trans_linear}
\alias{trans_mgks}
\alias{trans_nexpsm}
\alias{trans_linear_nexpsm}
\alias{trans_inter_linear}
\alias{trans_inter_nexp}
\alias{trans_inter_mgks}
\title{Specifying inner transformations}
\usage{
trans_linear(pord, S, alpha, a0)

trans_mgks(y0, alpha)

trans_nexpsm(S, alpha)

trans_linear_nexpsm(
  pord = NULL,
  S_si = NULL,
  S_nexp = NULL,
  alpha_nexp = NULL,
  alpha_si = NULL,
  alpha_scale = NULL,
  center = FALSE,
  alpha_center = NULL,
  n_si = NULL,
  n_nexp = NULL,
  Z0 = NULL,
  positive_si = FALSE
)

trans_inter_linear(pord = NULL, S = NULL, alpha = NULL, a0 = NULL)

trans_inter_nexp(pord = NULL, S = NULL, alpha = NULL)

trans_inter_mgks(pord = NULL, S = NULL, alpha = NULL)
}
\arguments{
\item{pord}{The order of the differences of the transformation parameters
that will be penalised. E.g., \code{pord = 1} corresponds to penalising the squared
first order differences \eqn{(alpha_k - alpha_{k-1})^2}. Ignored if argument \code{S}
is supplied.}

\item{S}{Positive semi-definite matrix used to specify a generalised ridge
penalty on the transformation parameters. In \code{trans_linear}, if \code{S} is not supplied, the \code{pord}-order differences
between parameters are penalised, see the \code{pord} argument.}

\item{alpha}{Vector containing the initial values for the parameters of the transformation. If provided they can be used
by the optimiser.}

\item{y0}{Vector of \eqn{n} observations corresponding to the rows of the variables in \code{s_nest}.}

\item{S_si}{Penalty matrix used in trans_linear_nexpsm. For the linear transformation parameters.}

\item{S_nexp}{Penalty matrix used in expsmooth. For the linear transformation parameters.}

\item{alpha_nexp}{Initial values used in trans_linear_nexpsm. For expsmooth parameters.}

\item{alpha_si}{Initial values used in trans_linear_nexpsm. For linear transformation parameters.}

\item{n_si}{Length of \code{alpha_si} if \code{alpha_si} is not provided.}

\item{n_nexp}{Length of \code{alpha_nexp} if \code{alpha_nexp} is not provided.}
}
\description{
Functions to specify nested covariate transformations, to be used
within \link{s_nest}.
}
\details{
The types of transformations currently provided are:
\itemize{
\item{\code{trans_linear}{ a linear transformation \eqn{X^\top \alpha}, which can be used to specify, e.g, a single index vector (i.e. a projection) or
                      a linear effect.}}
\item{\code{trans_mgks}{ a multivariate kernel smooth transformation based on the response variable observation 
                         vector \code{y0} and corresponding distance matrix.}}
\item{\code{trans_nexpsm}{ an exponential smoothing transformation.}}
\item{\code{trans_linear_nexpsm}{ a combination of a linear transformation and an exponential smoothing transformation.}}
}
}
