% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bundle_deriv.R
\name{wrap_bundle_deriv}
\alias{wrap_bundle_deriv}
\title{Check the derivatives of a margin bundle}
\usage{
wrap_bundle_deriv(bundle, X, trans = NULL, w = NULL)
}
\arguments{
\item{bundle}{a margin bundle, e.g. \code{bundle_exp()}.}

\item{X}{the term of the margin, as passed to \link{s_nest}.}

\item{trans}{optional arguments of the margin, e.g. \code{trans_linear(pord = 1)}.}

\item{w}{weights of the observations. Random by default: \eqn{\sum_i z_i} alone is always 0 because \eqn{z} is centred.}
}
\value{
A list with the functions \code{d0}, \code{d1}, \code{d2}, \code{d3} of the inner parameters, to be passed
to \link{check_deriv}, and \code{param}, the initial inner parameters set by the bundle.
}
\description{
Wraps the \code{eval} function of a margin bundle (\link{nest_bundles}) so that its derivatives can
be compared with finite differences by \link{check_deriv}, as done for a family with the derivatives of
\link{llk_gaussian}. The function differentiated is \eqn{\sum_i w_i z_i}, where \eqn{z} is the index of
the margin computed by the bundle at the term \code{X}.
}
\examples{
# see ?nest_bundles
}
